+++ lib/phoenix_kit_ecommerce/web/user_order_details.ex
@@ -52,7 +52,7 @@ defmodule PhoenixKitEcommerce.Web.UserOrderDetails do
         if order.user_uuid != current_user.uuid do
           {:ok,
            socket
-           |> put_flash(:error, Gettext.gettext(PhoenixKitWeb.Gettext, "Access denied"))
+           |> put_flash(:error, gettext("Access denied"))
            |> push_navigate(to: Routes.path("/dashboard/orders"))}
         else
           {:ok, setup_order_assigns(socket, order, current_user)}